Implement differential term
This uses a 150 second sliding buffer giving equal weight to stuff
that happened 150 seconds ago as more recently in an attempt to
smooth out the very sharp edges of temperature sensor steps.

But I think maybe when the temperature is settled, this is
introducing some slight oscillation. Maybe it can be smoothed in
a different way.